Move LanguageFactory in tests
