+++ backend/blog/tests/factories.py
@@ -3,7 +3,7 @@ import factory.fuzzy
 from blog.models import Post
 from django.utils import timezone
 from factory.django import DjangoModelFactory
-from i18n.helpers.tests import LanguageFactory
+from i18n.tests.factories import LanguageFactory
 from pytest_factoryboy import register
 from users.tests.factories import UserFactory
