+++ backend/hotels/tests/factories.py
@@ -3,7 +3,7 @@ import factory.fuzzy
 from conferences.tests.factories import ConferenceFactory
 from factory.django import DjangoModelFactory
 from hotels.models import HotelRoom
-from i18n.helpers.tests import LanguageFactory
+from i18n.tests.factories import LanguageFactory
 from pytest_factoryboy import register
